v1.0.0
Getting aframe-cubemap-component to work with the latest version of
A-Frame requires a breaking change. At some point, Three.js removed
the 'tCube' shader uniform which cubemap-component relied on. The fix
is described in this SO question: https://stackoverflow.com/a/59454999/6591491
This fix is apparently not compatible with older versions of
three.js/A-Frame (at least it does not work with A-Frame v0.7.0).